+++ common/src/main/java/com/tristankechlo/explorations/init/ModRegistry.java
@@ -12,6 +12,7 @@ import com.tristankechlo.explorations.worldgen.structures.UndergroundTempleStruc
 import com.tristankechlo.explorations.worldgen.structures.pieces.SlimeCaveStructurePiece;
 import com.tristankechlo.explorations.worldgen.structures.processors.DeepslateProcessor;
 import com.tristankechlo.explorations.worldgen.structures.processors.StoneBrickAgingProcessor;
+import com.tristankechlo.explorations.worldgen.structures.processors.WoolReplacingProcessor;
 import com.tristankechlo.explorations.worldgen.treedecorators.CaveVineDecorator;
 import com.tristankechlo.explorations.worldgen.treedecorators.LanternDecorator;
 import net.minecraft.core.Registry;
@@ -52,6 +53,7 @@ public final class ModRegistry {
     public static final RegistrationProvider<StructureProcessorType<?>> STRUCTURE_PROCESSORS = RegistrationProvider.get(Registry.STRUCTURE_PROCESSOR, Explorations.MOD_ID);
     public static final RegistryObject<StructureProcessorType<DeepslateProcessor>> DEEPSLATE_PROCESSOR = STRUCTURE_PROCESSORS.register("deepslate", () -> () -> DeepslateProcessor.CODEC);
     public static final RegistryObject<StructureProcessorType<StoneBrickAgingProcessor>> STONE_BRICK_AGING_PROCESSOR = STRUCTURE_PROCESSORS.register("stone_brick_aging", () -> () -> StoneBrickAgingProcessor.CODEC);
+    public static final RegistryObject<StructureProcessorType<WoolReplacingProcessor>> WOOL_REPLACING_PROCESSOR = STRUCTURE_PROCESSORS.register("wool_replacing", () -> () -> WoolReplacingProcessor.CODEC);
 
     /* CONFIGURED FEATURES */
     public static final RegistrationProvider<ConfiguredFeature<?, ?>> CONFIGURED_FEATURES = RegistrationProvider.get(BuiltinRegistries.CONFIGURED_FEATURE, Explorations.MOD_ID);